moved under sourmash_lib; removed prime; fixed tests
