+++ .gitignore
@@ -15,3 +15,4 @@ sourmash.egg-info
 .ipynb_checkpoints
 _minhash.so
 .cache
+*.so
