+++ Makefile
@@ -7,4 +7,4 @@ clean:
 	$(PYTHON) setup.py clean --all
 
 test: all
-	$(PYTHON) -m pytest sourmash_lib.py sourmash_signature.py test__minhash.py
+	$(PYTHON) -m pytest sourmash_lib/*.py
