+++ setup.py
@@ -21,11 +21,12 @@ SETUP_METADATA = \
     "author": "C. Titus Brown",
     "author_email": "titus@idyll.org",
     "license": "BSD 3-clause",
-    "py_modules": ["sourmash_lib","sourmash_signature"],
-    "ext_modules": [Extension("_minhash",
-                              sources=["_minhash.cc",
+    "packages": ["sourmash_lib"],
+    "ext_modules": [Extension("sourmash_lib._minhash",
+                              sources=["sourmash_lib/_minhash.cc",
                                        "third-party/smhasher/MurmurHash3.cc"],
-                              depends=["_minhash.hh", "kmer_min_hash.hh"],
+                              depends=["sourmash_lib/_minhash.hh",
+                                       "sourmash_lib/kmer_min_hash.hh"],
                               language="c++",
                               extra_compile_args=EXTRA_COMPILE_ARGS)],
     "scripts": ["sourmash"],
